refactor(core): Adapt namespaces
