+++ src/Http/Controllers/Discord/DiscordOAuthController.php
@@ -27,10 +27,9 @@ namespace Herpaderpaldent\Seat\SeatNotifications\Http\Controllers\Discord;
 
 use Exception;
 use GuzzleHttp\Client;
-use Herpaderpaldent\Seat\SeatNotifications\Actions\SubscribeAction;
+use Herpaderpaldent\Seat\SeatNotifications\Http\Actions\SubscribeAction;
 use Herpaderpaldent\Seat\SeatNotifications\Caches\RedisRateLimitProvider;
 use Herpaderpaldent\Seat\SeatNotifications\Models\Discord\DiscordUser;
-use Illuminate\Http\Request;
 use RestCord\DiscordClient;
 use RestCord\Model\User\User;
 use UnexpectedValueException;
